+++ src/graphnet/models/__init__.py
@@ -1,1 +1,2 @@
 from .model import Model
+from .simple_model import SimpleModel
